// src/api/orval/model/createOrUpdateQuotaPlanRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { QuotaOverlimitAction } from "./quotaOverlimitAction";

export interface CreateOrUpdateQuotaPlanRequest {
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  id?: number;
  /**
   * @minLength 1
   * @maxLength 64
   */
  name: string;
  /**
   * @maxLength 256
   * @nullable
   */
  description?: string | null;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @nullable
   */
  dailyQuotaBytes?: number | null;
  /**
   * @minimum 0
   * @maximum 9223372036854776000
   * @nullable
   */
  monthlyQuotaBytes?: number | null;
  /**
   * @minimum 0
   * @maximum 2147483647
   * @nullable
   */
  upKbps?: number | null;
  /**
   * @minimum 0
   * @maximum 2147483647
   * @nullable
   */
  downKbps?: number | null;
  overlimitAction: QuotaOverlimitAction;
  /**
   * @minimum 0
   * @maximum 2147483647
   * @nullable
   */
  throttleUpKbps?: number | null;
  /**
   * @minimum 0
   * @maximum 2147483647
   * @nullable
   */
  throttleDownKbps?: number | null;
  isActive?: boolean;
  isDefault?: boolean;
}