// src/api/orval/model/quotaPlanDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { QuotaOverlimitAction } from "./quotaOverlimitAction";

export interface QuotaPlanDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dailyQuotaBytes?: number | null;
  /** @nullable */
  monthlyQuotaBytes?: number | null;
  /** @nullable */
  upKbps?: number | null;
  /** @nullable */
  downKbps?: number | null;
  overlimitAction?: QuotaOverlimitAction;
  /** @nullable */
  throttleUpKbps?: number | null;
  /** @nullable */
  throttleDownKbps?: number | null;
  isActive?: boolean;
  isDefault?: boolean;
}