// src/api/orval/model/quotaOverlimitAction.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export type QuotaOverlimitAction =
  (typeof QuotaOverlimitAction)[keyof typeof QuotaOverlimitAction];

export const QuotaOverlimitAction = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
} as const;